refactor: memory accesses inside functiondef node
